Added Travis support.
We need to consider the case of a PR and a branch separately.

The main problem is that Travis only "load" the local branch, so it
doesn't know things likes origin/master. Also in a PR, Travis call the
branch master so we need to find the correct branch.

There is only one case when this doesn't work and is if you run git-cop
rspec test on Travis. But as Circle CI is used this shouldn't be a
problem.

Fixes https://github.com/bkuhlmann/git-cop/issues/12