Fixed Travis CI pull request build hook.
The code to support Travis from
c3c46bcda162770cd7cdc8bc222557fb9ac0f617
was refactored in
1d3e4e71c30728cf2b14bc6e191676e1516d335b

The refactored code missed the following line:
`git fetch original_branch test_name:test_name`
It is needed because when creating a PR Travis doesn't know the
original branch as it sets it to `master`, except if the PR come from
the same project and not from a fork of it. Due to this Travis worked
only in local branchs and in PR comming from the same project.

This may be a Travis bug, but it doesn't hurt anyway.